Add option to specify timezone for a specific date
This option is needed because of the previous patch. For a specific date
we need to know what timezone is needed. Some country has different
timezone according to the date in the year. For now, the user must
specify the timezone. Later we will try to introduce a feature which
integrate timezone automatically.